refactor(dashboard): extract ConnectionRow + ModelCompatPopover + SiliconFlowEndpointModal — #3501 Phase 1d
#3501 Phase 1d: god-component 8,092->6,838 LOC.
